+++ nebula/config.go
@@ -218,5 +218,5 @@ type configFirewallRule struct {
 type configRelay struct {
 	AmRelay   bool     `yaml:"am_relay,omitempty"`
 	UseRelays bool     `yaml:"use_relays"`
-	relays    []string `yaml:"relays,omitempty"`
+	Relays    []string `yaml:"relays,omitempty"`
 }